% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-10-cat-alternation.R
\name{alternate}
\alias{alternate}
\title{categorical vectors can 'alternate' between their basic levels and alternative values for each level.
the functions in this file all relate to accessing, switching, changing etc. these alternatives
Set categorical vector to alternative values}
\usage{
alternate(x, alternative = c(), internal = FALSE)
}
\arguments{
\item{x}{categorical vector (see [categorical()])}

\item{alternative}{the alternative value as a string}

\item{internal}{logical: set to TRUE if you want to alternate to an internal alternative (useful for vector classes that are based on the categorical class)}
}
\value{
the original vector, but its active values are replaced by the alternative
}
\description{
categorical vectors can 'alternate' between their basic levels and alternative values for each level.
the functions in this file all relate to accessing, switching, changing etc. these alternatives
Set categorical vector to alternative values
}
