% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_categorical.R
\name{as_categorical}
\alias{as_categorical}
\title{Create a categorical variable}
\usage{
as_categorical(x, ..., classification)
}
\arguments{
\item{x}{variable to which to add th classification}

\item{...}{when \code{classification} is missing \code{\link{as_categorical}}
is called with these arguments to create the classification}

\item{classification}{classification to add to \code{x}}
}
\description{
Create a categorical variable
}
