% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_classification.R
\name{as_classification}
\alias{as_classification}
\title{Convert a data.frame to a classification object}
\usage{
as_classification(x, compute_parent = FALSE, order = 1:4)
}
\arguments{
\item{x}{the data.frame to convert. See, details for the format of the
data.frame}

\item{compute_parent}{compute the parent column of the input}

\item{order}{selection of columns from x.}
}
\description{
Convert a data.frame to a classification object
}
\details{
The data.frame should contain the following columns:
\describe{
  \item{id}{the id of the category (cast to character)}
  \item{label}{the label of the category (cast to character)}
  \item{level}{the level of the classification in which the category belongs
    (should be integer)}
  \item{parent}{the id of the parent category (cast to character). Can be
    omitted when \code{compute_parent = FALSE}. Should contain missing values
    for categories in level 1 of the classification.}
}
}
