% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_classification_simple.R
\name{as_classification_simple}
\alias{as_classification_simple}
\title{Create classification with only one level}
\usage{
as_classification_simple(id, label)
}
\arguments{
\item{id}{vector of category ids}

\item{label}{vector of category labels}
}
\description{
Create classification with only one level
}
