% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-11-cat_methods.R
\name{cat_remove_levels}
\alias{cat_remove_levels}
\title{remove levels from a categorical vector}
\usage{
cat_remove_levels(x, choices)
}
\arguments{
\item{x}{categorical vector}

\item{choices}{levels to be removed}
}
\description{
remove levels from a categorical vector
}
