% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{categorical.cat_categorical}
\alias{categorical.cat_categorical}
\title{create a categorical variable from categorical input}
\usage{
\method{categorical}{cat_categorical}(x = logical(), levels = NULL,
  alternatives = NULL, alternatives_internal = NULL,
  active_alternative = NULL, active_alternative_is_internal = FALSE,
  class = c())
}
\description{
create a categorical variable from categorical input
}
