% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{categorical.categorical}
\alias{categorical.categorical}
\title{create a categorical variable from categorical input}
\usage{
\method{categorical}{categorical}(x = logical(), levels = levels(x),
  alternatives_internal = NULL, alternatives = NULL,
  active_alternative = NULL, active_alternative_is_internal = FALSE,
  class = c())
}
\description{
create a categorical variable from categorical input
}
