% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{categorical.matrix}
\alias{categorical.matrix}
\title{categorical constructors check inputs, convert to matrix and finally should call this function.}
\usage{
\method{categorical}{matrix}(x = logical(), levels,
  alternatives_internal = tibble::tibble(.rows = length(levels)), ...,
  class = c())
}
\description{
categorical constructors check inputs, convert to matrix and finally should call this function.
}
