% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-00-cat.R
\name{categorical_logic}
\alias{categorical_logic}
\title{mutate categorical type variables, while treating each choice as logical}
\usage{
categorical_logic(x, ...)
}
\arguments{
\item{...}{arguments passed to dplyr::mutate}

\item{.data}{a data.frame or tibble}
}
\value{
see ?dplyr::mutate
}
\description{
This is much simpler than it sounds & useful; needs better description & name
}
\details{
operates rowwise (see ?dplyr::rowwise) on a categorical column. Each row's value is a vector with the selected responses.
}
