% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_classification.R
\name{complete_classification}
\alias{complete_classification}
\title{Complete classification in case branches in tree are not all of same depth}
\usage{
complete_classification(x, order = 1:4)
}
\arguments{
\item{x}{the data.frame to convert. See, details for the format of the
data.frame}

\item{order}{selection of columns from x.}
}
\description{
Complete classification in case branches in tree are not all of same depth
}
\details{
Some classification do not all go to the same depth for each of the branches. 
The function \code{\link{as_classification}} does not support that. This 
function repeats categories so that all branches are of the same depth.
}
