% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-12-cat_utilities.R
\name{enforce_alternative_lengths_match_levels}
\alias{enforce_alternative_lengths_match_levels}
\title{ensures that the provided alternative values have as many rows as there are levels}
\usage{
enforce_alternative_lengths_match_levels(alternatives, levels)
}
\arguments{
\item{alternatives}{a tibble from attributes(some_categorical_vector)$alternatives}

\item{levels}{character vector with the corresponding levels}
}
\description{
ensures that the provided alternative values have as many rows as there are levels
}
