% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_data.R
\docType{data}
\name{example_choices}
\alias{example_choices}
\title{Example ODK questionnaire choices}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1119 rows and 11 columns.}
\source{
\url{http://www.diamondse.info/}
}
\usage{
example_choices
}
\description{
An odk survey consists of two sheets; the questions and the choices. This is the sheet with the list of choices as a data.frame.
Matching the two other example files 'example_questions' and 'example_data'.
}
\keyword{datasets}
