% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data set: Needs assessment in Somalia in June 1018}
\format{An object of class \code{list} of length 925.}
\source{
\url{http://www.diamondse.info/}
}
\usage{
example_data
}
\description{
The example is a small subset of the original stratified sample and should not be used in practice. Contact the package maintainer for the full dataset
Matching the two other example files 'example_questions' and 'example_choices' (which are the odk xlsform as data.frames.).
}
\keyword{datasets}
