% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_data.R
\docType{data}
\name{example_questions}
\alias{example_questions}
\title{Example ODK questionnaire survey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 506 rows and 14 columns.}
\source{
\url{http://www.diamondse.info/}
}
\usage{
example_questions
}
\description{
An odk survey consists of two sheets; the questions and the choices. This is the sheet with the list of questions as a data.frame.
Matching the two other example files 'example_choices' and 'example_data'.
}
\keyword{datasets}
