% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_multiple_logical_operators.R
\name{\%==exactly\%}
\alias{\%==exactly\%}
\title{Logical operators for select multiple}
\usage{
x \%==exactly\% y
}
\arguments{
\item{x}{values to check}

\item{y}{values to check against}
}
\description{
Logical operators for select multiple
}
