% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_range.R
\name{interval}
\alias{interval}
\title{create a new interval variable}
\usage{
interval(x, levels = unique(unlist(x)), ranks = 1:length(levels), ...)
}
\arguments{
\item{x}{a vector of to be used as values for the interval vector. These should be characters for most use cases (but can be other types)}

\item{levels}{vector of of possible values for x; similar to factor levels}

\item{...}{named vectors with alternative values corresponding to 'levels'. Must each have the same length as levels. Can be accessed with \code{alternate}. These "external" alternatives are open to user defined alternatives, for example labels in multiple languages.}

\item{rank}{a vector of numeric ranks corresponding to each level.}
}
\description{
create a new interval variable
}
