% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{is.categorical}
\alias{is.categorical}
\title{check if vector is of class cat_categorical}
\usage{
is.categorical(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if it is a categorical vector
}
\description{
check if vector is of class cat_categorical
}
