% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-00-range.R
\name{is.interval}
\alias{is.interval}
\title{check if vector is of class cat_interval}
\usage{
is.interval(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if it is a categorical vector
}
\description{
check if vector is of class cat_interval
}
