% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-00-ordinal.R
\name{is.ordinal}
\alias{is.ordinal}
\title{check if vector is of class cat_ordinal}
\usage{
is.ordinal(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if it is a categorical vector
}
\description{
check if vector is of class cat_ordinal
}
