% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_multiple_s3.R
\name{is.select_multiple}
\alias{is.select_multiple}
\title{check if vector is of class cat_select_multiple}
\usage{
is.select_multiple(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if it is
}
\description{
check if vector is of class cat_select_multiple
}
