% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{is_alternated}
\alias{is_alternated}
\title{is a categorical vector alternated?}
\usage{
is_alternated(x)
}
\arguments{
\item{x}{a vector of type categorical}
}
\value{
logical, TRUE if the categorical vector is alternated. For details see \code{\link{alternate}}
}
\description{
is a categorical vector alternated?
}
