% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_ordinal.R
\name{is_ordinal}
\alias{is_ordinal}
\title{check if vector is of class cat_ordinal}
\usage{
is_ordinal(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if it is a categorical vector
}
\description{
check if vector is of class cat_ordinal
}
