% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-04-cat-casting-utilities.R
\name{join_levels}
\alias{join_levels}
\title{combine the levels attribute of two categorical vectors}
\usage{
join_levels(x, y)
}
\arguments{
\item{x}{first categorical vector}

\item{y}{second categorical vector}
}
\description{
combine the levels attribute of two categorical vectors
}
