% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label.odk_select_multiple}
\alias{label.odk_select_multiple}
\title{labels for select multiple}
\usage{
\method{label}{odk_select_multiple}(x)
}
\arguments{
\item{x}{a list of class categorical_select_multiple}
}
\value{
same as `x`, but all values are replaced by values; lookup table for labels no longer part of attributes.
}
\description{
labels for select multiple
}
