% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{apply labels to categorical data}
\usage{
label(x)
}
\arguments{
\item{x}{a categorical vector of class categorical_select_multiple or categorical_select_one}
}
\value{
a character vector with the labels of `x`
}
\description{
apply labels to categorical data
}
\seealso{
\code{\link{select_multiple}}
}
