% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_multiple_s3.R
\name{mutate_select_multiple}
\alias{mutate_select_multiple}
\title{Mutate select_multiple type variables in a data frame}
\usage{
mutate_select_multiple(.data, ...)
}
\arguments{
\item{.data}{a data.frame or tibble}

\item{...}{arguments passed to dplyr::mutate}
}
\value{
see ?dplyr::mutate
}
\description{
Mutate select_multiple type variables in a data frame
}
\details{
operates rowwise (see ?dplyr::rowwise) on a select_multiple column. Each row's value is a factor vector with the selected responses.
}
