% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nace_rev2}
\alias{nace_rev2}
\title{NACE Rev.2 Classification}
\format{An object of class \code{classification} of length 4.}
\source{
\url{http://ec.europa.eu/eurostat/ramon/nomenclatures/index.cfm?TargetUrl=LST_CLS_DLD&StrNom=NACE_REV2&StrLanguage#}
}
\usage{
nace_rev2
}
\description{
Contains the Statistical Classification of Economic Activities in the 
European Community, Rev. 2 in level_meta format.
}
\keyword{datasets}
