% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-00-cat.R
\name{new_categorical}
\alias{new_categorical}
\title{create a new categorical variable}
\usage{
new_categorical(x = logical(), levels,
  alternatives_internal = empty_alternatives(levels),
  alternatives = empty_alternatives(levels), active_alternative = NULL,
  active_alternative_is_internal = FALSE, class = c())
}
\arguments{
\item{x}{a logical matrix indicating which levels are selected per record (each row is a record, each column corresponds to a level specified in 'level's)}

\item{levels}{list of possible values for x; similar to factor levels}

\item{alternatives_internal}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{alternate}. "internal" alternatives are used to store 'fixed' alternatives for classes extending 'cat_categorical'.}

\item{alternatives}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{\link{alternate}}.}

\item{active_alternative}{alternative to be used for display}

\item{active_alternative_is_internal}{whether the active alternative is internal}

\item{class}{optional: name the subclass if you are using this to define a new vector type on top of the 'cat_categorical' class}
}
\description{
create a new categorical variable
}
