% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk_select_multiple_s3.R
\name{new_select_multiple}
\alias{new_select_multiple}
\title{create a new select_multiple variable}
\usage{
new_select_multiple(x = character(), choices = NULL, labels = NULL,
  sep = " ")
}
\arguments{
\item{x}{a character vector with concatenated select_multiple choices (for example 'c("choice_A choices_B", "choice_C")`)}

\item{choices}{list of options; equivalent to factor levels (in case some options were never selected but we want to track them regardless)}

\item{labels}{named vector with choice labels. the vector name is the value in `x`, the vector value is the label.}

\item{sep}{the delimeter used to separate the choices in each element of `x` ("choice_A choice_B" vs. "choice_A; choice_B"). uses regex.}
}
\description{
create a new select_multiple variable
}
