% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-00-ordinal.R
\name{ordinal}
\alias{ordinal}
\title{create a new ordinal variable}
\usage{
ordinal(x, levels = unique(unlist(x)), ranks = 1:length(levels), ...)
}
\arguments{
\item{x}{a vector of to be used as values for the ordinal vector. These should be characters for most use cases (but can be other types)}

\item{levels}{vector of of possible values for x; similar to factor levels}

\item{ranks}{a vector of numeric ranks corresponding to each level.}

\item{...}{named vectors with alternative values corresponding to 'levels'. Must each have the same length as levels. Can be accessed with \code{alternate}. These "external" alternatives are open to user defined alternatives, for example labels in multiple languages.}
}
\description{
create a new ordinal variable
}
