% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-13-cat_plot.R
\name{plot.cat_categorical}
\alias{plot.cat_categorical}
\title{plotting categorical vectors}
\usage{
\method{plot}{cat_categorical}(x)
}
\arguments{
\item{x}{categorical vector}
}
\description{
plotting categorical vectors
}
