% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_classification}
\alias{read_classification}
\title{Read classification from file.}
\usage{
read_classification(filename, binary = TRUE)
}
\arguments{
\item{filename}{the file to read from.}

\item{binary}{is \code{TRUE} interpret the file as a binary (RDS) file. 
Otherwise the file is interpreted as a CSV file.}
}
\description{
Read classification from file.
}
