% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-08-cat_access_attributes.R
\name{restore_lgl_list_NA_in_value_list}
\alias{restore_lgl_list_NA_in_value_list}
\title{set a list of items to NA where any value in a categorical logical matrix representation is NA}
\usage{
restore_lgl_list_NA_in_value_list(value_list, x_categorical)
}
\arguments{
\item{value_list}{a list with as many items as there are records in x_categorical}

\item{x_categorical}{a categorical vector}
}
\description{
set a list of items to NA where any value in a categorical logical matrix representation is NA
}
