% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_classification}
\alias{save_classification}
\title{Save classification to file}
\usage{
save_classification(classification, filename, binary = TRUE)
}
\arguments{
\item{classification}{object of type \code{classification}.}

\item{filename}{the filename of the file to save to. When the extension is
'.csv' or '.CSV' the classification is saved as a CSV file.}

\item{binary}{save as a binary (RDS) file. Otherwise save as a CSV file.}
}
\description{
Save classification to file
}
