% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-10-cat-alternation.R
\name{set_active_alternative}
\alias{set_active_alternative}
\title{set a new alternative as active}
\usage{
set_active_alternative(x, alternative = character(), internal = FALSE)
}
\arguments{
\item{x}{categorical vector}

\item{alternative}{the name of the alternative to be set as the active one}

\item{internal}{logical, TRUE if the active alternative is internal}
}
\description{
set a new alternative as active
}
