% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{superficial_nas}
\alias{superficial_nas}
\title{find superficial NAs}
\usage{
superficial_nas(x)
}
\arguments{
\item{x}{a <categorical> vectors}
}
\description{
find superficial NAs
}
\details{
"superficial NA's" appear in categorical vectors where the levels themselves are not NA, but the active alternative has no value for the level
}
