% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{unique_and_not_na}
\alias{unique_and_not_na}
\title{take unique values from a vector and remove all NAs}
\usage{
unique_and_not_na(x)
}
\arguments{
\item{x}{vector}
}
\description{
take unique values from a vector and remove all NAs
}
