% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/00-03-cat-casting-cat_categorical_cat_categorical.R
\name{vec_cast.cat_categorical.cat_categorical}
\alias{vec_cast.cat_categorical.cat_categorical}
\title{cast cat_categorical to cat_categorical}
\usage{
\method{vec_cast.cat_categorical}{cat_categorical}(x, to, ...)
}
\description{
cast cat_categorical to cat_categorical
}
