% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-02-cat-casting_character.R
\name{vec_cast.cat_categorical.character}
\alias{vec_cast.cat_categorical.character}
\title{cast character to cat_categorical}
\usage{
\method{vec_cast.cat_categorical}{character}(x, to, ...)
}
\arguments{
\item{x}{vector}

\item{to}{prototype to conver to}

\item{...}{additional arguments}
}
\description{
cast character to cat_categorical
}
