% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-01-cat-casting-boilerplates.R
\name{vec_cast.cat_categorical.default}
\alias{vec_cast.cat_categorical.default}
\title{cast categorical default}
\usage{
\method{vec_cast}{cat_categorical.default}(x, to, ...)
}
\arguments{
\item{x}{vector}

\item{to}{prototype to conver to}

\item{...}{additional arguments}
}
\description{
cast categorical default
}
