% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-02-cat-casting_character.R
\name{vec_cast.character.cat_categorical}
\alias{vec_cast.character.cat_categorical}
\title{cast cat_categorical to character}
\usage{
\method{vec_cast.character}{cat_categorical}(x, to, ...)
}
\description{
cast cat_categorical to character
}
