% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-06-cat_coersion.R
\name{vec_ptype.cat_categorical}
\alias{vec_ptype.cat_categorical}
\title{When the change happens implicitly (e.g in c()) we call it coercion
vec_ptype2(x, y) defines possible set of coercions.
It returns a prototype if x and y can be safely coerced to the same prototype;
otherwise it returns an error.
The set of automatic coercions is usually quite small
because too many tend to make code harder to reason about
and silently propagate mistakes.}
\usage{
\method{vec_ptype}{cat_categorical}(x)
}
\description{
each combo of our class and other classes that can be coerced together has
a function that returns just a prototype of the resulting class
}
\details{
namespace exports:
generic double dispatch boilerplate like this:
#' @method vec_cast CLASS
#' @export
#' @export vec_cast.CLASS
vec_cast.vctrs_percent <- function(x, to, ...) {
}

individual dispatches like this:

#' @method vec_cast.CLASS1 CLASS2
#' @export
cat_categorical
}
