% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/00-03-cat-casting-cat_categorical_cat_categorical.R
\name{vec_ptype2.cat_categorical.cat_categorical}
\alias{vec_ptype2.cat_categorical.cat_categorical}
\title{common type of cat_categorical and cat_categorical}
\usage{
\method{vec_ptype2.cat_categorical}{cat_categorical}(x, y, ...)
}
\description{
common type of cat_categorical and cat_categorical
}
