% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-02-cat-casting_character.R
\name{vec_ptype2.cat_categorical.character}
\alias{vec_ptype2.cat_categorical.character}
\title{the functions in this file determine how categorical vectors are converted to other vector types
this follows the vctrs package casting system (see ?browseVignettes("vctrs") for details)
common type: cat_categorical & character}
\usage{
\method{vec_ptype2.cat_categorical}{character}(x, y, ...)
}
\description{
the functions in this file determine how categorical vectors are converted to other vector types
this follows the vctrs package casting system (see ?browseVignettes("vctrs") for details)
common type: cat_categorical & character
}
