% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-01-cat-casting-boilerplates.R
\name{vec_ptype2.cat_categorical.default}
\alias{vec_ptype2.cat_categorical.default}
\title{default common type for cat_categorical}
\usage{
\method{vec_ptype2}{cat_categorical.default}(x, y, ..., x_arg = "x",
  y_arg = "y")
}
\arguments{
\item{x}{vector object}

\item{y}{vector object}

\item{...}{additional arguments}

\item{x_arg}{x argument}

\item{y_arg}{y argument}
}
\description{
default common type for cat_categorical
}
