% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-02-cat-casting_character.R
\name{vec_ptype2.character.cat_categorical}
\alias{vec_ptype2.character.cat_categorical}
\title{common type: character & cat_categorical}
\usage{
\method{vec_ptype2.character}{cat_categorical}(x, y, ...)
}
\arguments{
\item{x}{vector object}

\item{y}{vector object}

\item{...}{additional arguments}
}
\description{
common type: character & cat_categorical
}
