% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{alternate}
\alias{alternate}
\title{Set categorical vector to alternative vales}
\usage{
alternate(x, alternative = character(), internal = FALSE)
}
\arguments{
\item{x}{categorical vector (see \link{\code{categorical()}})}

\item{alternative}{the alternative value as a string}
}
\value{
the original vector, but its active values are replaced by the alternative
}
\description{
Set categorical vector to alternative vales
}
