% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-00-cat.R
\name{as_categorical}
\alias{as_categorical}
\title{convert to categorical variable}
\usage{
as_categorical(x = logical(), levels = NULL,
  alternatives = empty_alternatives(levels),
  alternatives_internal = empty_alternatives(levels),
  active_alternative = NULL, active_alternative_is_internal = FALSE,
  class = c())
}
\arguments{
\item{x}{a vector or list to be used as values for the categorical vector}

\item{levels}{list of possible values for x; similar to factor levels}

\item{alternatives}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{\link{alternate}}.}

\item{alternatives_internal}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{alternate}. "internal" alternatives are used to store 'fixed' alternatives for classes extending 'cat_categorical'.}

\item{active_alternative}{alternative to be used for display}

\item{active_alternative_is_internal}{whether the active alternative is internal}

\item{class}{optional: name the subclass if you are using this to define a new vector type on top of the 'cat_categorical' class}
}
\description{
convert to categorical variable
}
