% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{as_categorical}
\alias{as_categorical}
\title{create a new categorical variable}
\usage{
as_categorical(x = logical(), levels = unique(unlist(x)),
  alternatives_internal = tibble::tibble(.rows = length(levels)), ...)
}
\arguments{
\item{x}{a vector or list to be used as values for the categorical vector}

\item{levels}{list of possible values for x; similar to factor levels}

\item{alternatives_internal}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{categorical_alternative}. "internal" alternatives are used to store 'fixed' alternatives for classes extending 'cat_categorical'.}

\item{...}{named vectors with alternative values corresponding to 'levels'. Must each have the same length as levels. Can be accessed with \code{categorical_alternative}. These "external" alternatives are open to user defined alternatives, for example labels in multiple languages.}
}
\description{
create a new categorical variable
}
