% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical_base_methods.R
\name{cat_remove_levels}
\alias{cat_remove_levels}
\title{remove levels from a categorical vector}
\usage{
cat_remove_levels(x, choices)
}
\description{
remove levels from a categorical vector
}
