% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{categorical.categorical}
\alias{categorical.categorical}
\title{create a categorical variable from categorical input}
\usage{
\method{categorical}{categorical}(x = logical(),
  levels = unique_and_not_na(x),
  alternatives_internal = tibble::tibble(.rows = length(levels)), ...,
  class = c())
}
\description{
create a categorical variable from categorical input
}
