% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-00-cat.R
\name{categorical.matrix}
\alias{categorical.matrix}
\title{categorical constructors check inputs, convert to matrix and finally should call this function.}
\usage{
\method{categorical}{matrix}(x = logical(), levels,
  alternatives = empty_alternatives(levels),
  alternatives_internal = empty_alternatives(levels),
  active_alternative = NULL, active_alternative_is_internal = FALSE,
  class = c())
}
\description{
categorical constructors check inputs, convert to matrix and finally should call this function.
}
