% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-00-cat.R
\name{categorical.matrix}
\alias{categorical.matrix}
\title{categorical constructors check inputs, convert to matrix and finally should call this function.}
\usage{
\method{categorical}{matrix}(x = logical(), levels,
  alternatives = empty_alternatives(levels),
  alternatives_internal = empty_alternatives(levels), class = c())
}
\arguments{
\item{x}{a logical matrix; each row corresponding to a value in the resulting vector, each column to a level / category}

\item{levels}{list of possible values for x; similar to factor levels}

\item{alternatives}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{\link{alternate}}.}

\item{alternatives_internal}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{alternate}. "internal" alternatives are used to store 'fixed' alternatives for classes extending 'cat_categorical'.}

\item{class}{optional: name the subclass if you are using this to define a new vector type on top of the 'cat_categorical' class}
}
\description{
categorical constructors check inputs, convert to matrix and finally should call this function.
}
