% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{categorical}
\alias{categorical}
\title{cat_categorical is the meta class that gives the general structure for the more specific categorical classes that extend it.
cat_categorical is:
   - a vctrs_vctr
   - values are generally stored as a list (to allow select multiple and other more complex subclasses)
   - it has an attribute for levels / allowed values.
   - it has an attribute for _closed alternative values_. what these are depends on the specific sublass; for example these could be:
       - character labels (select)
       - integer rank (ordinal)
   - it has an attribute for _open alternative values_, allowing the user to add different alternative values, such as labels in different languages}
\usage{
categorical(x = logical(), levels = NULL,
  alternatives = empty_alternatives(levels),
  alternatives_internal = empty_alternatives(levels),
  active_alternative = NULL, active_alternative_is_internal = FALSE,
  class = c())
}
\arguments{
\item{x}{a vector or list to be used as values for the categorical vector}

\item{levels}{list of possible values for x; similar to factor levels}

\item{alternatives}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{\link{alternate}}.}

\item{alternatives_internal}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{alternate}. "internal" alternatives are used to store 'fixed' alternatives for classes extending 'cat_categorical'.}

\item{active_alternative}{alternative to be used for display}

\item{ative_alternative_is_internal}{whether the active alternative is internal}

\item{...}{named vectors with alternative values corresponding to 'levels'. Must each have the same length as levels. Can be accessed with \code{alternate}. These "external" alternatives are open to user defined alternatives, for example labels in multiple languages.}
}
\description{
create a new categorical variable
}
