% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-11-cat_methods.R
\name{droplevels.cat_categorical}
\alias{droplevels.cat_categorical}
\title{drop levels from categorical vector}
\usage{
\method{droplevels}{cat_categorical}(x, ...)
}
\description{
drop levels from categorical vector
}
