% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-10-cat-alternation.R
\name{list_alternatives}
\alias{list_alternatives}
\title{List all alternative values for a categorical vector}
\usage{
list_alternatives(x, internal = NULL)
}
\arguments{
\item{x}{a categorical vector}

\item{internal}{logical: If TRUE, show internal alternatives only}
}
\value{
a list with internal and public alternatives as character vectors, or only one of them as a vector if `internal` is set
}
\description{
List all alternative values for a categorical vector
}
