% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical.R
\name{new_categorical}
\alias{new_categorical}
\title{create a new categorical variable}
\usage{
new_categorical(x = logical(), levels,
  alternatives_internal = empty_alternatives(levels),
  alternatives = empty_alternatives(levels), active_alternative = NULL,
  active_alternative_is_internal = FALSE, class = c())
}
\arguments{
\item{x}{a logical matrix indicating which levels are selected per record (each row is a record, each column corresponds to a level specified in 'level's)}

\item{levels}{vector of possible values for x; similar to factor levels. Defaults to the unique values in x. Will be converted to characters}

\item{alternatives_internal}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{categorical_alternative}. "internal" alternatives are used to store 'fixed' alternatives for classes extending 'cat_categorical'.}

\item{alternatives}{a named list of vectors with alternative values corresponding to 'levels'. Must have the same length as levels. Can be accessed with \code{categorical_alternative}. These "external" alternatives are open to user defined alternatives, for example labels in multiple languages.}
}
\description{
create a new categorical variable
}
