% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-00-range.R
\name{new_interval}
\alias{new_interval}
\title{create a new interval variable (EXPERIMENTAL)}
\usage{
new_interval(lower, upper, closed = c(TRUE, FALSE), levels, levels_lower,
  levels_upper, ...)
}
\arguments{
\item{lower}{numeric vector of lower limits}

\item{upper}{numeric vector of upper limits}

\item{closed}{logical vector of length 2: are lower or upper ends open?}

\item{levels}{vector of of possible values for x; similar to factor levels}

\item{levels_lower}{?}

\item{levels_upper}{?}

\item{...}{named vectors with alternative values corresponding to 'levels'. Must each have the same length as levels. Can be accessed with \code{alternate}. These "external" alternatives are open to user defined alternatives, for example labels in multiple languages.}
}
\description{
create a new interval variable (EXPERIMENTAL)
}
