% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-10-cat-alternation.R
\name{set_active_alternative}
\alias{set_active_alternative}
\title{set a new alternative as active}
\usage{
set_active_alternative(x, alternative = character(), internal = FALSE)
}
\description{
set a new alternative as active
}
