% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical_cast_boilerplates.R
\name{vec_cast.cat_categorical}
\alias{vec_cast.cat_categorical}
\title{cast categorical vectors}
\usage{
\method{vec_cast}{cat_categorical}(x, to, ...)
}
\arguments{
\item{x}{vector}

\item{to}{prototype to conver to}

\item{...}{additional arguments}
}
\description{
cast categorical vectors
}
