% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_categorical_coersion.R
\name{vec_ptype2.cat_categorical}
\alias{vec_ptype2.cat_categorical}
\title{BOILERPLATES}
\usage{
\method{vec_ptype2}{cat_categorical}(x, y, ...)
}
\arguments{
\item{x}{vector object}

\item{y}{vector object}

\item{...}{additional arguments}
}
\description{
start with the boilerplate for vec_ptype2()
this is just so later we can write vec_ptype2.cat_categorical.OTHERCLASS
to define how these two classes should be coerced together
}
